refactor: migrate JSON library from yason to jzon
This commit migrates the JSON library from yason to jzon for the
following reasons:

1. Easier passing of vector (unsigned byte 8) or stream to jzon for
webhook construction:
- Most people use clack as their web server abstraction layer, thus
it's easier to pass in the payload when constructing webhooks.

2. Improved performance with hash tables instead of plists:
- Destructuring a hash table is generally quicker than a plist for
large JSON objects with nested structures.
- Switching from plists to hash tables allows for quicker access to
elements.

This commit includes the following changes:
- Replace yason usage with jzon for parsing.
- Update the initialize methods for objects to work with hash table
instead of plist including vectors of hash tables when we have a list
of objects.
- Specify the correct return type for a list of objects from
`:type list` to `:type vector`.